修改了H264Analysis的代码风格

+++ H264AnalysisQt/h264analysisqt.h
@@ -28,8 +28,8 @@ public:
 		QTableWidgetItem *naluSizeItem;
 	} NaluItem, *PNaluItem;
 	typedef struct {
-		size_t pos;
-		size_t size;
+		UINT32 pos;
+		UINT32 size;
 		QString str;
 	} Nalu, *PNalu;
 public:
